lambda-workana/requirements.txt
playwright==1.45.0
boto3
requests
python-dotenv